major code cleanup, some type changes in API, bump version to 2.1.0

+++ src/snc/gen_ss_code.c
@@ -167,7 +167,7 @@ void gen_ss_code(Program *program)
 			/* Generate event processing function */
 			gen_state_func(ssp->value, sp->value,
 				sp->state_whens, gen_event_body,
-				"Event", "E", "long",
+				"Event", "E", "unsigned",
 				", short *pTransNum, short *pNextState");
 			/* Generate action processing function */
 			gen_state_func(ssp->value, sp->value,
